@@ -20,23 +20,23 @@
         "sort-packages": true
     },
     "extra": {
-        "branch-alias": {
-            "dev-master": "1.1-dev"
-        },
         "laminas": {
             "module": "StatusLib"
         }
     },
     "require": {
-        "php": "^5.6 || ^7.0",
-        "laminas-api-tools/api-tools-configuration": "^1.0",
-        "laminas/laminas-config": "^3.1",
-        "laminas/laminas-db": "^2.8.2",
-        "laminas/laminas-hydrator": "^1.1 || ^2.0 || ^3.0",
-        "laminas/laminas-paginator": "^2.7",
+        "php": "^7.3 || ~8.0.0",
+        "laminas-api-tools/api-tools-configuration": "^1.4",
+        "laminas/laminas-config": "^3.4",
+        "laminas/laminas-db": "^2.12",
+        "laminas/laminas-hydrator": "^1.1 || ^2.0 || ^3.0 || ^4.0",
+        "laminas/laminas-paginator": "^2.10",
         "laminas/laminas-stdlib": "^2.7.7 || ^3.0.1",
         "laminas/laminas-zendframework-bridge": "^1.0",
         "ramsey/uuid": "^2.8"
+    },
+    "require-dev": {
+        "laminas/laminas-coding-standard": "~2.3.0"
     },
     "autoload": {
         "psr-4": {
@@ -45,5 +45,12 @@
     },
     "replace": {
         "zfcampus/statuslib-example": "^1.2.0"
+    },
+    "scripts": {
+        "check": [
+            "@cs-check"
+        ],
+        "cs-check": "phpcs --colors",
+        "cs-fix": "phpcbf --colors"
     }
 }
