@@ -1,19 +1,15 @@
 <?php
 
-/**
- * @see       https://github.com/laminas-api-tools/statuslib-example for the canonical source repository
- * @copyright https://github.com/laminas-api-tools/statuslib-example/blob/master/COPYRIGHT.md
- * @license   https://github.com/laminas-api-tools/statuslib-example/blob/master/LICENSE.md New BSD License
- */
+declare(strict_types=1);
 
 namespace StatusLib;
 
 return [
-    'statuslib' => [
+    'statuslib'       => [
         // 'array_mapper_path' => 'path/to/PHP/file/returning/array.php',
     ],
     'service_manager' => [
-        'aliases' => [
+        'aliases'   => [
             Mapper::class => ArrayMapper::class,
 
             // Legacy Zend Framework aliases
