@@ -1,16 +1,18 @@
 <?php
 
-/**
- * @see       https://github.com/laminas-api-tools/statuslib-example for the canonical source repository
- * @copyright https://github.com/laminas-api-tools/statuslib-example/blob/master/COPYRIGHT.md
- * @license   https://github.com/laminas-api-tools/statuslib-example/blob/master/LICENSE.md New BSD License
- */
+declare(strict_types=1);
 
 namespace StatusLib;
 
 use DomainException;
 use Laminas\ApiTools\Configuration\ConfigResource;
 use Laminas\Config\Writer\PhpArray as ConfigWriter;
+use Psr\Container\ContainerInterface;
+
+use function file_exists;
+use function is_array;
+use function realpath;
+use function sprintf;
 
 /**
  * Service factory for the ArrayMapper
@@ -25,6 +27,10 @@ use Laminas\Config\Writer\PhpArray as ConfigWriter;
  */
 class ArrayMapperFactory
 {
+    /**
+     * @param ContainerInterface $services
+     * @return ArrayMapper
+     */
     public function __invoke($services)
     {
         if (! $services->has('config')) {
