@@ -1,12 +1,11 @@
 <?php
 
-/**
- * @see       https://github.com/laminas-api-tools/statuslib-example for the canonical source repository
- * @copyright https://github.com/laminas-api-tools/statuslib-example/blob/master/COPYRIGHT.md
- * @license   https://github.com/laminas-api-tools/statuslib-example/blob/master/LICENSE.md New BSD License
- */
+declare(strict_types=1);
 
 namespace StatusLib;
+
+use stdClass;
+use Traversable;
 
 /**
  * Interface for StatusLib mappers
@@ -14,13 +13,13 @@ namespace StatusLib;
 interface MapperInterface
 {
     /**
-     * @param array|\Traversable|\stdClass $data 
+     * @param array|Traversable|stdClass $data
      * @return Entity
      */
     public function create($data);
 
     /**
-     * @param string $id 
+     * @param string $id
      * @return Entity
      */
     public function fetch($id);
@@ -31,14 +30,14 @@ interface MapperInterface
     public function fetchAll();
 
     /**
-     * @param string $id 
-     * @param array|\Traversable|\stdClass $data 
+     * @param string $id
+     * @param array|Traversable|stdClass $data
      * @return Entity
      */
     public function update($id, $data);
 
     /**
-     * @param string $id 
+     * @param string $id
      * @return bool
      */
     public function delete($id);
